--- src/main/kotlin/blackjack/model/Card.kt
@@ -1,3 +1,0 @@
-package blackjack.model
-
-data class Card(val shape: CardShape, val number: CardNumber)
