--- src/main/kotlin/blackjack/model/CardDeck.kt
@@ -1,10 +1,0 @@
-package blackjack.model
-
-object CardDeck {
-    val cardDeck =
-        CardShape.entries.flatMap { shape ->
-            CardNumber.entries.map { number ->
-                Card(shape, number)
-            }
-        }.shuffled()
-}
