--- src/main/kotlin/blackjack/model/CardGenerator.kt
@@ -1,5 +1,0 @@
-package blackjack.model
-
-interface CardGenerator {
-    fun draw(): Card
-}
