--- src/main/kotlin/blackjack/model/CardHandState.kt
@@ -1,10 +1,0 @@
-package blackjack.model
-
-enum class CardHandState(
-    val precondition: Int = 0,
-) {
-    BLACKJACK(21),
-    BUST,
-    STAY,
-    HIT,
-}
