--- src/main/kotlin/blackjack/model/CardNumber.kt
@@ -1,19 +1,0 @@
-package blackjack.model
-
-enum class CardNumber(
-    val number: Int,
-) {
-    ACE(11),
-    TWO(2),
-    THREE(3),
-    FOUR(4),
-    FIVE(5),
-    SIX(6),
-    SEVEN(7),
-    EIGHT(8),
-    NINE(9),
-    TEN(10),
-    KING(10),
-    QUEEN(10),
-    JACK(10),
-}
