--- src/main/kotlin/blackjack/model/CardShape.kt
@@ -1,8 +1,0 @@
-package blackjack.model
-
-enum class CardShape {
-    HEART,
-    CLOVER,
-    SPADE,
-    DIAMOND,
-}
