--- src/main/kotlin/blackjack/model/DealerWinning.kt
@@ -1,3 +1,0 @@
-package blackjack.model
-
-data class DealerWinning(val result: Map<WinningResultStatus, Int>)
