--- src/main/kotlin/blackjack/model/Player.kt
@@ -1,13 +1,0 @@
-package blackjack.model
-
-data class Player(override val name: PlayerName, override val cardHand: CardHand) : Role(name, cardHand) {
-    override fun getState(): CardHandState {
-        val sum = cardHand.sum()
-
-        return when {
-            sum > CardHandState.BLACKJACK.precondition -> CardHandState.BUST
-            sum == CardHandState.BLACKJACK.precondition -> CardHandState.BLACKJACK
-            else -> CardHandState.STAY
-        }
-    }
-}
