--- src/main/kotlin/blackjack/model/PlayerWinning.kt
@@ -1,3 +1,0 @@
-package blackjack.model
-
-data class PlayerWinning(val result: Map<PlayerName, WinningResultStatus>)
