--- src/main/kotlin/blackjack/model/WinningResultStatus.kt
@@ -1,15 +1,0 @@
-package blackjack.model
-
-enum class WinningResultStatus(val output: String) {
-    VICTORY("승"),
-    DEFEAT("패"),
-    DRAW("무"),
-    ;
-
-    fun reverse(): WinningResultStatus =
-        when (this) {
-            VICTORY -> DEFEAT
-            DEFEAT -> VICTORY
-            DRAW -> DRAW
-        }
-}
