+++ src/main/kotlin/blackjack/model/card/Card.kt
@@ -1,0 +1,3 @@
+package blackjack.model.card
+
+data class Card(val shape: CardShape, val number: CardNumber)
