+++ src/main/kotlin/blackjack/model/card/CardDeck.kt
@@ -1,0 +1,10 @@
+package blackjack.model.card
+
+object CardDeck {
+    val cardDeck =
+        CardShape.entries.flatMap { shape ->
+            CardNumber.entries.map { number ->
+                Card(shape, number)
+            }
+        }.shuffled()
+}
