+++ src/main/kotlin/blackjack/model/card/CardShape.kt
@@ -1,0 +1,8 @@
+package blackjack.model.card
+
+enum class CardShape {
+    HEART,
+    CLOVER,
+    SPADE,
+    DIAMOND,
+}
