+++ src/main/kotlin/blackjack/model/card/generator/CardGenerator.kt
@@ -1,0 +1,7 @@
+package blackjack.model.card.generator
+
+import blackjack.model.card.Card
+
+interface CardGenerator {
+    fun draw(): Card
+}
