+++ src/main/kotlin/blackjack/model/playing/cardhand/CardHandState.kt
@@ -1,0 +1,10 @@
+package blackjack.model.playing.cardhand
+
+enum class CardHandState(
+    val precondition: Int = 0,
+) {
+    BLACKJACK(21),
+    BUST,
+    STAY,
+    HIT,
+}
