+++ src/main/kotlin/blackjack/model/winning/DealerWinning.kt
@@ -1,0 +1,3 @@
+package blackjack.model.winning
+
+data class DealerWinning(val result: Map<WinningResultStatus, Int>)
