+++ src/main/kotlin/blackjack/model/winning/PlayerWinning.kt
@@ -1,0 +1,5 @@
+package blackjack.model.winning
+
+import blackjack.model.playing.participants.player.PlayerName
+
+data class PlayerWinning(val result: Map<PlayerName, WinningResultStatus>)
